Moves all Types to a single file. The previous layout looked like overengineered mess :P
