+++ pasta.cabal
@@ -17,10 +17,7 @@ library
   hs-source-dirs:      src
   ghc-options:         -Wall
   exposed-modules:     Pasta
-  other-modules:       Pasta.Prelude
-                     , Pasta.Insert
-                     , Pasta.Update
-                     , Pasta.Select
+  other-modules:       Pasta.Types
   build-depends:       base >= 4.7 && < 5
                      , lens >= 4.13
                      , text
